consolidate and format item variant form
